Added more files
So I added my head to the <li> section. I need help with getting the CSS in the right place and still am just confused. I know I will get this. 
Im also not sure if my top bladee image should be linked in the index, the css, or both. I also want to center the list in the index, but think this should be a css function.